Address both of sinbads concerns to an extent.
