+++ redbot/cogs/customcom/customcom.py
@@ -14,7 +14,7 @@ from redbot.core import Config, checks, commands
 from redbot.core.i18n import Translator, cog_i18n
 from redbot.core.utils import AsyncIter
 from redbot.core.utils.chat_formatting import box, pagify, escape, humanize_list
-from redbot.core.utils.menus import SimpleHybridMenu
+from redbot.core.utils._dpy_menus_utils import SimpleHybridMenu
 from redbot.core.utils.predicates import MessagePredicate
 from .menus import CCRawSource, CCListSource
