+++ redbot/cogs/economy/menus.py
@@ -9,7 +9,7 @@ from redbot.core import bank
 from redbot.core.i18n import Translator
 from redbot.core.utils.chat_formatting import humanize_number, box, escape
 
-from redbot.core.utils.menus import SimpleHybridMenu
+from redbot.core.utils._dpy_menus_utils import SimpleHybridMenu
 from redbot.vendored.discord.ext import menus
 
 _ = Translator("Economy", __file__)
