+++ redbot/cogs/modlog/menus.py
@@ -8,7 +8,7 @@ from redbot.core.i18n import Translator
 from redbot.core.modlog import Case
 from redbot.core.utils.chat_formatting import humanize_number
 
-from redbot.core.utils.menus import SimpleHybridMenu
+from redbot.core.utils._dpy_menus_utils import SimpleHybridMenu
 from redbot.vendored.discord.ext import menus
 
 _ = Translator("General", __file__)
