+++ redbot/cogs/modlog/modlog.py
@@ -9,7 +9,7 @@ from redbot.core import checks, commands, modlog
 from redbot.core.bot import Red
 from redbot.core.i18n import Translator, cog_i18n
 from redbot.core.utils.chat_formatting import box
-from redbot.core.utils.menus import SimpleHybridMenu
+from redbot.core.utils._dpy_menus_utils import SimpleHybridMenu
 from redbot.core.utils.predicates import MessagePredicate
 
 from .menus import CasesForSource
